requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
scikit-image>=0.21
scipy>=1.11
lime>=0.2
umap-learn>=0.5
Pillow>=9.0
